test: move vm-threads to cli
