+++ test/cli/fixtures/vm-threads/module.test.js
