+++ test/cli/fixtures/vm-threads/src/external/css/processed.module.css
