+++ test/cli/fixtures/vm-threads/src/external/default-cjs.cjs
