+++ test/cli/fixtures/vm-threads/src/external/export-default-cjs.js
