+++ test/cli/fixtures/vm-threads/src/external/export-nested-default-cjs.js
