+++ test/cli/fixtures/vm-threads/src/external/nested-default-cjs.cjs
