+++ test/cli/fixtures/vm-threads/src/external/not-found.js
