+++ test/cli/fixtures/vm-threads/src/external/package-null/package-null.js
