+++ test/cli/fixtures/vm-threads/src/external/package.json
