+++ test/cli/fixtures/vm-threads/src/external/primitive-cjs.cjs
