--- test/vm-threads/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "@vitest/test-vm-threads",
-  "type": "module",
-  "private": true,
-  "scripts": {
-    "test": "vitest"
-  },
-  "devDependencies": {
-    "jsdom": "latest",
-    "vitest": "workspace:*"
-  }
-}
